+++ src/binDataFile.h
@@ -35,24 +35,24 @@ class binDataFile : public binData {
 		bool isOpen();
 		bool isEOF();
 		
-		int seek(u_int32_t ulOffset);																					//Move the pointer to a file offset
-		u_int32_t offset();																								//Return the current offset
+		int seek(u_int64_t ulOffset);																					//Move the pointer to a file offset
+		u_int64_t offset();																								//Return the current offset
 		
-		int getData(void* pData, u_int32_t ulSize, u_int32_t* p_ulSizeRead);								//Read ulSize bytes of data
-		int getData(void* pData, u_int32_t ulSize, u_int32_t ulOffset, u_int32_t* p_ulSizeRead);	//Read ulSize bytes of data from offset ulOffset
+		int getData(void* pData, u_int64_t ulSize, u_int64_t* p_ulSizeRead);								//Read ulSize bytes of data
+		int getData(void* pData, u_int64_t ulSize, u_int64_t ulOffset, u_int64_t* p_ulSizeRead);	//Read ulSize bytes of data from offset ulOffset
 		
-		int skipNullBlocks(u_int8_t cByteWidth, u_int32_t* p_posNew);										// From the current position in the file, skip over
+		int skipNullBlocks(u_int8_t cByteWidth, u_int64_t* p_posNew);										// From the current position in the file, skip over
 																																// cByteWidth blocks of NULL data. (Useful for sparse
 																																// files.)
 
 		//int getData(vector<char>* pData, u_int32_t ulSize, u_int32_t* p_ulReadSize = NULL);		//Read up to ulSize bytes of data and report actual bytes read into p_ulReadSize
 		//int getData(vector<char>* pData, u_int32_t ulSize, u_int32_t ulOffset, u_int32_t* p_ulReadSize = NULL);	//Read up to ulSize bytes of data from offset ulOffset and report actual bytes read into p_ulReadSize
 		
-		int getString(string* pString, u_int32_t ulLength);													//Read ulLength characters into pString
-		int getString(string* pString, u_int32_t ulOffset, u_int32_t ulLength);							//Read ulLength characters from offset ulOffset into pString
+		int getString(string* pString, u_int64_t ulLength);													//Read ulLength characters into pString
+		int getString(string* pString, u_int64_t ulOffset, u_int64_t ulLength);							//Read ulLength characters from offset ulOffset into pString
 
-		int getTwoByteCharString(string* pString, u_int32_t ulLength, bool bBigEndian=false);
-		int getTwoByteCharString(string* pString, u_int32_t ulOffset, u_int32_t ulLength, bool bBigEndian=false);
+		int getTwoByteCharString(string* pString, u_int64_t ulLength, bool bBigEndian=false);
+		int getTwoByteCharString(string* pString, u_int64_t ulOffset, u_int64_t ulLength, bool bBigEndian=false);
 				
 	private:
 		ifstream m_filestream;
